_gcc_48_compat
Gbp-Pq: Name 0013_gcc_48_compat.patch

+++ test/cctest/test-macro-assembler-x64.cc
@@ -2185,7 +2185,7 @@ TEST(PositiveSmiTimesPowerOfTwoToInteger64) {
 TEST(OperandOffset) {
   v8::internal::V8::Initialize(NULL);
   int data[256];
-  for (int i = 0; i < 256; i++) { data[i] = i * 0x01010101; }
+  for (int i = 0; i < 256; i++) { data[i] = (long)i * 0x01010101; }
 
   // Allocate an executable page of memory.
   size_t actual_size;
